// frontend/src/enums/meteor-enums.ts
export enum eMeteorType {
    STONY = "Stony",
    IRON = "Iron",
    STONY_IRON = "Stony-Iron",
}

export enum eMeteorMaterial {
    STONE = "Stone",
    IRON = "Iron",
    MIXED = "Mixed"
}

export enum eWeatherType {
    CLEAR = "Clear",
    RAIN = "Rain",
    SNOW = "Snow",
    STORM = "Storm"
}